data_processing/requirements.txt
annotated-types==0.5.0
azure-core==1.29.2
azure-cosmos==4.5.0
azure-eventhub==5.11.4
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.1
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pathy==0.10.2
preshed==3.0.8
pydantic==2.1.1
pydantic_core==2.4.0
regex==2023.8.8
requests==2.31.0
six==1.16.0
smart-open==6.3.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
thinc==8.1.12
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wasabi==1.1.2
prometheus_client==0.17.1